+++ app/src/main/java/com/dam/spacereporter/spacereporter/ui/launches/LaunchesFragment.java
@@ -1,4 +1,4 @@
-package com.dam.spacereporter.spacereporter.ui.fragments;
+package com.dam.spacereporter.spacereporter.ui.launches;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
