+++ app/src/main/res/layout/fragment_launches.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".spacereporter.ui.fragments.LaunchesFragment">
+    tools:context=".spacereporter.ui.launches.LaunchesFragment">
 
     <!-- TODO Implement UI -->
     <TextView
