+++ app/src/main/res/layout/fragment_read_later.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".spacereporter.ui.fragments.ReadLaterFragment">
+    tools:context=".spacereporter.ui.readlater.ReadLaterFragment">
 
     <!-- TODO Implement UI -->
     <TextView
