+++ app/src/main/res/layout/fragment_settings.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context=".spacereporter.ui.fragments.SettingsFragment">
+    tools:context=".spacereporter.ui.settings.SettingsFragment">
 
     <!-- TODO Implement UI -->
